syn_back_flask/requirements.txt
alembic==1.7.5
click==8.0.3
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.10.0
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
SQLAlchemy==1.4.28
Werkzeug==2.0.2
zipp==3.6.0
flask-marshmallow==0.14.0
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
matplotlib==3.5.1
mypy-extensions==0.4.3
numpy==1.22.3
packaging==21.3
pathspec==0.9.0
paho-mqtt==1.6.1
Shapely==1.8.1.post1
redis==4.4.2